requirements.txt
aioredis==2.0.1
amqp==5.1.1
asgi-redis==1.4.3
asgiref==3.4.1
async-timeout==4.0.3
attrs==23.1.0
autobahn
Automat==22.10.0
billiard>=4.2.0,<5.0
botocore==1.31.39
celery==5.4.0rc2
certifi==2023.7.22
cffi==1.15.1
channels==3.0.4
channels-redis==4.1.0
charset-normalizer==3.2.0
click>=8.1.2,<9.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.3
daphne==3.0.2
Django==3.2.18
django-celery-results==2.2.0
future==0.18.3
hiredis==2.2.3
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iso8601==2.0.0
jmespath==1.0.1
kombu>=5.3.4,<6.0
macaroonbakery==1.3.1
msgpack==1.0.5
msgpack-python==0.5.6
networkx==3.1
oauthlib==3.2.2
prompt-toolkit==3.0.39
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyRFC3339==1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
requests-unixsocket==0.3.0
rsa==4.9
s3transfer==0.6.2
service-identity==23.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing-extensions==4.12.2
tzdata==2023.3
urllib3==1.26.16
vine>=5.1.0,<6.0
wcwidth==0.2.6
zope.interface==6.0
